requirements.txt
numpy
pandas
scikit-learn
torch
torchvision
Pillow
opencv-python